% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_imputation_functions.R
\name{impute.MF}
\alias{impute.MF}
\title{Imputation of Missing Protein Abundances using MissForest}
\usage{
impute.MF(data = as.matrix(data), maxiter_MF, ntree, maxnodes)
}
\arguments{
\item{data}{dataset in the form of a matrix or data frame with either NAs or 0s as missing values}

\item{maxiter_MF}{maximum number of iteration to be performed if the stopping criteria is not met beforehand}

\item{ntree}{number of trees to grow in each forest}

\item{maxnodes}{maximum number of terminal nodes for trees in the forest, has to equal at least the number of columns in the given data}
}
\value{
the imputed version of the dataset
}
\description{
The function impute.MF imputes a dataset with missing values or NA's using MissForest
}
\examples{
#' #' \dontrun{
data(datapnnl)
data<-datapnnl.rm.ref[1:100,1:21]
impute.MF(data=as.matrix(data), maxiter_MF = 10, ntree = 100,maxnodes = NULL)
}
}
