% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegImpute.R
\name{returnTargets}
\alias{returnTargets}
\title{Return target set for current column.}
\usage{
returnTargets(col, data, present_indices)
}
\arguments{
\item{col}{A number indicating current columns index.}

\item{data}{A matrix, The original data matrix with NaNs.}

\item{present_indices}{A list containing indices in current column where values are not missing.}
}
\value{
A list of length n.
}
\description{
Return target set for current column.
}
