% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegImpute.R
\name{returnTestSet}
\alias{returnTestSet}
\title{Return back filled data for test indices.}
\usage{
returnTestSet(col, filled, test_indices)
}
\arguments{
\item{col}{A number indicating current columns index.}

\item{test_indices}{A list of indices corresponding to cells in current column that contain NaN.}
}
\value{
A matrix
}
\description{
Return back filled data for test indices.
}
